fix(resources/gta): fix `.as<uint32_t>` causing function call to crash
- This will crash if the number is negative (which it can often be in Lua)
- msgpack assumes if you pass `uint32_t` you only want *specifically* positive integers, it will not try to do any wrapping itself.

+++ code/components/citizen-resources-gta/src/PopulationCreationFunctions.cpp
@@ -41,7 +41,7 @@ static InitFunction initFunction([]()
 				}
 				else
 				{
-					state->model = args[0].as<uint32_t>();
+					state->model = static_cast<uint32_t>(args[0].as<int32_t>());
 				}
 			}
 		});
